% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_power.R
\name{calculate_power}
\alias{calculate_power}
\title{Calculating Powers in Non-Inferiority Trials with Treatment Switching}
\usage{
calculate_power(
  n,
  r = 1,
  m1 = 1,
  m2 = 1.5,
  shape = 1,
  f1 = NULL,
  m0 = 0.5,
  f2 = 0.75,
  margin = NULL,
  p.s = 0.3,
  r.s = 0.3,
  rho.s = 0.3,
  s.dist = "gamma",
  censoring.rate = c("AC.only"),
  Ta = 1.5,
  Te = 3,
  tau = NULL,
  one.sided.alpha = 0.025,
  TXswitch = c("1to2", "2to1")[1],
  n_simulations = 5000,
  seed = 2024
)
}
\arguments{
\item{n}{Number of participants in active control.}

\item{r}{The ratio of participants in the new treatment group to
those in the active control group. The default ratio is 1.}

\item{m1}{Median survival time in the active control group.}

\item{m2}{Median survival time in the new treatment group.}

\item{shape}{The shape parameter of Weibull distributions for the event times
of two groups. The shape parameters are assumed to be the same.}

\item{f1}{Preserved fraction (0 < f1 <= 1) for the RMST of active control.
If f1 = 0.8, RMST of new treatment is not lower than 80% RMST of active control.
That is, H1: RMST of new treatment > 0.8 RMST of active control, equivalent to
H1: (RMST of new treatment - RMST of active control) >
- (1 - f1)(RMST of active control). margin = (1 - f1)(RMST of active control).}

\item{m0}{Median survival time in placebo group.}

\item{f2}{Preserved fraction (0 < f2 <= 1) of the efficacy of active control compared to placebo.
H1: (RMST of new treatment - RMST of RMST of placebo) >
f2 (RMST of active control - RMST of placebo), equivalent to
H1: (RMST of new treatment - RMST of active control) >
- (1 - f2)(RMST of active control - RMST of placebo).
margin = (1 - f2)(RMST of active control - RMST of placebo).}

\item{margin}{A margin (>0) for DRMST.}

\item{p.s}{Proportion of patients who may switch from active control to
new treatment after evaluation. The default is 0.3}

\item{r.s}{The ratio of the mean switching time to the mean survival time in the
active control group. The default is 0.3.}

\item{rho.s}{Correlation between switching time and event time in active control.}

\item{s.dist}{An option for the distribution of switching times.
s.dist="gamma", "beta", "unif", "indepExp", or a constant value.}

\item{censoring.rate}{Censoring rate in active control under no treatment switching.
The default is c("AC.only").}

\item{Ta}{Accrual duration.}

\item{Te}{Trial duration.}

\item{tau}{A value to specify the truncation time point for the RMST calculation.
The default is NULL, denoting tau = Te.}

\item{one.sided.alpha}{One-sided significance level. The default is 0.025.}

\item{TXswitch}{direction of treatment switching (TXswitch="1to2" or "2to1").
The default TXswitch = "1to2".}

\item{n_simulations}{Number of simulations. The default is 5000.}

\item{seed}{set.seed(seed).}
}
\value{
\item{power}{The proportion of rejecting null hypothesis}

\item{E1}{The mean number of events in active control group}

\item{E2}{The mean number of events in new treatment group}

\item{mRMST1}{The mean DRMST value across simulations in the active control group}

\item{mRMST2}{The mean DRMST value across simulations in the new treatment group}

\item{mDRMST}{The mean DRMST value across simulations}

\item{margin}{margin}
}
\description{
A function to calculate powers in non-inferiority trials with treatment switching
by comparing restricted mean survival time between two groups
}
\examples{
# calculate_power(141, r=1, m1=1, m2=1.1, f1=0.8, p.s=0.3, tau=2.5)
# calculate_power(141, r=1, m1=1, m2=1.1, m0=0.5, f2=0.5, p.s=0.3, tau=2.5)

# mar <- margin.HR2DRMST(m1=1, shape=1, tau=2.5, theta=0.833)
# calculate_power(141, r=1, m1=1, m2=1.1, margin=mar$margin, p.s=0.3, tau=2.5)

# ab <- paramWeibull(m1=1, t1=2.5, surv.rate=0.1)
# calculate_power(141, r=1, m1=1, m2=1.1, shape=ab[1],
#                 m0=0.5, f2=0.5, p.s=0.3, tau=2.5)
}
