% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramGamma.R
\name{paramGamma}
\alias{paramGamma}
\title{Calculating Shape and Scale Parameters for Gamma Distribution}
\usage{
paramGamma(m, t1, surv.prob)
}
\arguments{
\item{m}{Median survival time.}

\item{t1}{Time t1.}

\item{surv.prob}{Survival probability at time t1.}
}
\description{
A function to calculate the shape (k) and scale parameters of a gamma distribution
given a median survival and a survival probability at a specified time.
}
\examples{
# paramGamma(m=1, t1=2.5, surv.prob=0.1)
}
