% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramWeibull.R
\name{paramWeibull}
\alias{paramWeibull}
\title{Calculating Shape and Scale Parameters}
\usage{
paramWeibull(m1, t1, surv.rate)
}
\arguments{
\item{m1}{Median survival time.}

\item{t1}{Time t1.}

\item{surv.rate}{What is the survival rate at time t1?}
}
\description{
A function to calculate the shape and scale parameters of a Weibull distribution
given a median survival and a survival rate at a specified time.
}
\examples{
# paramWeibull(m1=1, t1=2.5, surv.rate=0.1)
}
