% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramWeibull.R
\name{paramWeibull}
\alias{paramWeibull}
\title{Calculating Shape and Scale Parameters for Weibull Distribution}
\usage{
paramWeibull(m, t1, surv.prob)
}
\arguments{
\item{m}{Median survival time.}

\item{t1}{Time t1.}

\item{surv.prob}{What is the survival probability at time t1?}
}
\description{
A function to calculate the shape and scale parameters of a Weibull distribution
given a median survival and a survival probability at a specified time.
}
\examples{
# paramWeibull(m=1, t1=2.5, surv.prob=0.1)
}
